requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
#yPDE2
chromadb
